+++ homeassistant/components/airvisual/manifest.json
@@ -4,6 +4,5 @@
   "config_flow": true,
   "documentation": "https://www.home-assistant.io/integrations/airvisual",
   "requirements": ["pyairvisual==4.3.0"],
-  "dependencies": ["air_quality"],
   "codeowners": ["@bachya"]
 }
